+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- Added [Convert-CharCase][] the corresponding delimiter with the -AsSentence switch.
-- Added [Convert-CharCase][] parameters -ExcludePattern, -MatchPattern.
-- Added [Replace-InQuote][] -FirstMatch option.
+- Added [Convert-CharCase][] the corresponding delimiter with the `-AsSentence` switch.
+- Added [Convert-CharCase][] parameters `-ExcludePattern`, `-MatchPattern`.
+- Added [Replace-InQuote][] `-FirstMatch` option.
 - Added [Process-InQuote][] function.
 - Added [Process-CsvColumn][] function.
+- Added [Invoke-Link][] `-Push` option.
 
 ## [0.8.0] - 2025-04-27